Modularize i18n
