+++ common/i18n/consts.ts
@@ -1,0 +1,1 @@
+export const DEFAULT_LANGUAGE = "en";
